Professional Graphic Designer
Hi! My name is Emma. I'm an illustrator. Illustration is my Profession, so I would like to share that feeling with you. I specialize in Children's illustration and it is my most favorite kind of jobs. I believe that this type of art offers the opportunity to demonstrate creativity, imagination and creating my own original and fun characters, situations and environment is undoubtedly the best part of my job.